[SELC-7490] feat: add new section in add user form for EA + refactor to reduce logical complexity.
