+++ src/pages/addUserFlow/addUser/__tests__/AddUserForm.test.tsx
@@ -1,10 +1,13 @@
 import { fireEvent, screen, waitFor } from '@testing-library/react';
 import { RoleEnum } from '../../../../api/generated/onboarding/UserDto';
 import { mockedParties } from '../../../../microcomponents/mock_dashboard/data/party';
-import { mockedPartyProducts, mockedProductRoles } from '../../../../microcomponents/mock_dashboard/data/product';
+import {
+  mockedPartyProducts,
+  mockedProductRoles,
+} from '../../../../microcomponents/mock_dashboard/data/product';
 import { productRoles2ProductRolesList } from '../../../../model/ProductRole';
 import { renderWithProviders } from '../../../../utils/test-utils';
-import AddUserForm from '../components/AddUserForm';
+import AddUserForm from '../components/AddUserForm/AddUserForm';
 
 describe('AddUserForm Component', () => {
   const defaultProps = {
@@ -190,4 +193,3 @@ describe('AddUserForm Component', () => {
     expect(screen.getByRole('button', { name: /Continua/i })).toBeDisabled();
   });
 });
-
